Refactor into multi-file structure and add resource size
