+++ contentcuration/contentcuration/tests/node_metadata/test_query.py
@@ -1,21 +1,24 @@
+from django.db.models.aggregates import Sum
 from le_utils.constants import content_kinds
 from le_utils.constants import roles
 
 from ..base import StudioTestCase
 from ..testdata import node
 from ..testdata import tree
+from contentcuration.node_metadata.annotations import CoachCount
+from contentcuration.node_metadata.annotations import DescendantCount
+from contentcuration.node_metadata.annotations import HasChanged
+from contentcuration.node_metadata.annotations import ResourceCount
+from contentcuration.node_metadata.annotations import ResourceSize
+from contentcuration.node_metadata.annotations import SortOrderMax
+from contentcuration.node_metadata.query import Metadata
+from contentcuration.serializers import ContentNodeSerializer
 from contentcuration.serializers import RootNodeSerializer
-from contentcuration.serializers import SimplifiedContentNodeSerializer
-from contentcuration.utils.nodes import CoachCount
-from contentcuration.utils.nodes import DescendantCount
-from contentcuration.utils.nodes import HasChanged
-from contentcuration.utils.nodes import Metadata
-from contentcuration.utils.nodes import ResourceCount
-from contentcuration.utils.nodes import SortOrderMax
 
 
 DESCENDANT_COUNT = 'descendant_count'
 RESOURCE_COUNT = 'resource_count'
+RESOURCE_SIZE = 'resource_size'
 COACH_COUNT = 'coach_count'
 MAX_SORT_ORDER = 'max_sort_order'
 HAS_CHANGED_DESCENDANT = 'has_changed_descendant'
@@ -30,9 +33,11 @@ class MetadataTest(StudioTestCase):
         }, parent=parent)
         coach_video.role_visibility = roles.COACH
         coach_video.save()
+        coach_video.refresh_from_db()
+        return coach_video
 
     def set_tree_changed(self, tree, changed):
-        tree.get_descendants().update(changed=changed)
+        tree.get_descendants(include_self=True).update(changed=changed)
         tree.changed = changed
         tree.save()
 
@@ -41,7 +46,7 @@ class MetadataTest(StudioTestCase):
         query = Metadata([topic_tree_node.pk])
 
         results = query.annotate(**{DESCENDANT_COUNT: DescendantCount()})
-        serialized = SimplifiedContentNodeSerializer(topic_tree_node).data
+        serialized = ContentNodeSerializer(topic_tree_node).data
 
         self.assertEqual(serialized.get('metadata').get('total_count'),
                          results.get(serialized.get('id')).get(DESCENDANT_COUNT))
@@ -52,11 +57,39 @@ class MetadataTest(StudioTestCase):
         query = Metadata([topic_tree_node.pk])
 
         results = query.annotate(**{RESOURCE_COUNT: ResourceCount()})
-        serialized = SimplifiedContentNodeSerializer(topic_tree_node).data
+        serialized = ContentNodeSerializer(topic_tree_node).data
 
         self.assertEqual(serialized.get('metadata').get('resource_count'),
                          results.get(serialized.get('id')).get(RESOURCE_COUNT))
         self.assertEqual(5, results.get(serialized.get('id')).get(RESOURCE_COUNT))
+
+    def test_resource_size__topic(self):
+        topic_tree_node = tree()
+        nested_topic = topic_tree_node.get_descendants().filter(kind=content_kinds.TOPIC).first()
+        self.create_coach_node(nested_topic)
+
+        query = Metadata([topic_tree_node.pk])
+        results = query.annotate(**{RESOURCE_SIZE: ResourceSize()})
+        serialized = ContentNodeSerializer(topic_tree_node).data
+
+        self.assertEqual(serialized.get('metadata').get('resource_size'),
+                         results.get(serialized.get('id')).get(RESOURCE_SIZE))
+        self.assertEqual(0, results.get(serialized.get('id')).get(RESOURCE_SIZE))
+
+    def test_resource_size__video(self):
+        topic_tree_node = tree()
+        nested_topic = topic_tree_node.get_descendants().filter(kind=content_kinds.TOPIC).first()
+        video_node = self.create_coach_node(nested_topic)
+
+        query = Metadata([video_node.pk])
+        results = query.annotate(**{RESOURCE_SIZE: ResourceSize()})
+        serialized = ContentNodeSerializer(video_node).data
+
+        size = video_node.files.aggregate(size=Sum('file_size')).get('size')
+
+        self.assertEqual(serialized.get('metadata').get('resource_size'),
+                         results.get(serialized.get('id')).get(RESOURCE_SIZE))
+        self.assertEqual(size, results.get(serialized.get('id')).get(RESOURCE_SIZE))
 
     def test_coach_count(self):
         topic_tree_node = tree()
@@ -66,7 +99,7 @@ class MetadataTest(StudioTestCase):
         query = Metadata([topic_tree_node.pk])
 
         results = query.annotate(**{COACH_COUNT: CoachCount()})
-        serialized = SimplifiedContentNodeSerializer(topic_tree_node).data
+        serialized = ContentNodeSerializer(topic_tree_node).data
 
         self.assertEqual(serialized.get('metadata').get('coach_count'),
                          results.get(serialized.get('id')).get(COACH_COUNT))
@@ -137,29 +170,43 @@ class MetadataTest(StudioTestCase):
         topic_tree_node2 = tree()
 
         topic_tree1_topics = topic_tree_node1.get_descendants().filter(kind=content_kinds.TOPIC)
-        self.create_coach_node(topic_tree1_topics.first())
+        video_node = self.create_coach_node(topic_tree1_topics.first())
         self.set_tree_changed(topic_tree_node1, False)
+        self.set_tree_changed(topic_tree_node2, False)
         topic_tree1_topics.last().delete()
 
-        query = Metadata([topic_tree_node1.pk, topic_tree_node2.pk], **{
+        query = Metadata([topic_tree_node1.pk, topic_tree_node2.pk, video_node.pk], **{
             DESCENDANT_COUNT: DescendantCount(),
             RESOURCE_COUNT: ResourceCount(),
+            RESOURCE_SIZE: ResourceSize(),
             COACH_COUNT: CoachCount(),
-            HAS_CHANGED_DESCENDANT: HasChanged(),
+            HAS_CHANGED_DESCENDANT: HasChanged(include_self=True),
             MAX_SORT_ORDER: SortOrderMax(),
         })
 
         topic_tree1_results = query.get(topic_tree_node1.pk)
         topic_tree2_results = query.get(topic_tree_node2.pk)
+        video_node_results = query.get(video_node.pk)
+        print(video_node_results)
 
         self.assertEqual(6, topic_tree1_results.get(DESCENDANT_COUNT))
         self.assertEqual(5, topic_tree1_results.get(RESOURCE_COUNT))
+        self.assertEqual(0, topic_tree1_results.get(RESOURCE_SIZE))
         self.assertEqual(1, topic_tree1_results.get(COACH_COUNT))
-        self.assertFalse(topic_tree1_results.get(HAS_CHANGED_DESCENDANT))
+        self.assertTrue(topic_tree1_results.get(HAS_CHANGED_DESCENDANT))
         self.assertEqual(1, topic_tree1_results.get(MAX_SORT_ORDER))
 
         self.assertEqual(7, topic_tree2_results.get(DESCENDANT_COUNT))
         self.assertEqual(5, topic_tree2_results.get(RESOURCE_COUNT))
+        self.assertEqual(0, topic_tree2_results.get(RESOURCE_SIZE))
         self.assertEqual(0, topic_tree2_results.get(COACH_COUNT))
-        self.assertTrue(topic_tree2_results.get(HAS_CHANGED_DESCENDANT))
+        self.assertFalse(topic_tree2_results.get(HAS_CHANGED_DESCENDANT))
         self.assertEqual(2, topic_tree2_results.get(MAX_SORT_ORDER))
+
+        self.assertEqual(1, video_node_results.get(DESCENDANT_COUNT))
+        self.assertEqual(1, video_node_results.get(RESOURCE_COUNT))
+        self.assertEqual(video_node.files.aggregate(size=Sum('file_size')).get('size'),
+                         video_node_results.get(RESOURCE_SIZE))
+        self.assertEqual(1, video_node_results.get(COACH_COUNT))
+        self.assertFalse(video_node_results.get(HAS_CHANGED_DESCENDANT))
+        self.assertEqual(video_node.sort_order, video_node_results.get(MAX_SORT_ORDER))
