+++ contentcuration/contentcuration/utils/nodes.py
@@ -6,30 +6,13 @@ import os
 import uuid
 
 from django.conf import settings
-from django.contrib.postgres.aggregates.general import BoolOr
 from django.core.exceptions import ObjectDoesNotExist
 from django.core.exceptions import ValidationError
 from django.core.files.storage import default_storage
 from django.db import transaction
-from django.db.models import BooleanField
-from django.db.models import IntegerField
 from django.db.models import Q
-from django.db.models.aggregates import Count
-from django.db.models.aggregates import Max
-from django.db.models.aggregates import Sum
-from django.db.models.expressions import Case
-from django.db.models.expressions import F
-from django.db.models.expressions import Value
-from django.db.models.expressions import When
-from django.db.models.functions import Coalesce
-from django.db.models.sql.constants import LOUTER
 from django.utils.translation import ugettext as _
-from django_cte import With
-from le_utils.constants import content_kinds
-from le_utils.constants import roles
 
-from contentcuration.db.models.expressions import BooleanComparison
-from contentcuration.db.models.expressions import WhenQ
 from contentcuration.models import AssessmentItem
 from contentcuration.models import Channel
 from contentcuration.models import ContentNode
@@ -395,348 +378,3 @@ def move_node(node, parent=None, sort_order=None, channel_id=None):
                 n.tags.add(t)
 
     return node
-
-
-class MetadataCTE(object):
-    columns = []
-
-    def __init__(self, node_pks):
-        """
-        :param node_pks: list of ContentNode ID's
-        """
-        self.node_pks = node_pks
-        self.cte = None
-
-    def add_columns(self, columns):
-        self.columns.extend(columns)
-
-    def get(self):
-        if self.cte is None:
-            self.cte = self.build()
-        return self.cte
-
-    def build(self):
-        raise NotImplementedError('Build method must create CTE')
-
-    def join(self, query):
-        raise NotImplementedError('Join method must join query with CTE')
-
-    @property
-    def col(self):
-        return self.get().col
-
-
-class TreeMetadataCTE(MetadataCTE):
-    columns = ['tree_id']
-
-    def build(self):
-        tree_ids = ContentNode.objects.filter(pk__in=self.node_pks).values('tree_id')
-        return With(
-            ContentNode.objects.filter(tree_id__in=tree_ids).values(*set(self.columns)),
-            name='tree_cte'
-        )
-
-    def join(self, query):
-        cte = self.get()
-        return cte.join(query, tree_id=cte.col.tree_id).with_cte(cte)
-
-
-class FileMetadataCTE(MetadataCTE):
-    columns = ['content_id']
-
-    def build(self):
-        nodes = ContentNode.objects.filter(pk__in=self.node_pks)\
-            .exclude(kind_id=content_kinds.TOPIC)
-        files = nodes.values(
-            'content_id',
-            file_size=F('files__file_size'),
-            checksum=F('files__checksum')
-        )
-        assessment_files = nodes.values(
-            'content_id',
-            file_size=F('assessment_items__files__file_size'),
-            checksum=F('assessment_items__files__checksum')
-        )
-
-        return With(files.union(assessment_files).values(*set(self.columns)), name='file_cte')
-
-    def join(self, query):
-        cte = self.get()
-        return cte.join(query, content_id=cte.col.content_id, _join_type=LOUTER).with_cte(cte)
-        # query = sub_cte.join(nodes, _join_type=LOUTER, content_id=sub_cte.col.content_id)\
-        #     .values(*set(self.columns))\
-        #     .annotate(resource_size=Coalesce(Sum('file_size'), Value(0)))
-        #
-        # return With(query.with_cte(sub_cte))
-
-
-class MetadataAnnotation(object):
-    cte = None
-    cte_columns = ()
-
-    def get_annotation(self, cte):
-        """
-        :type cte: With|None
-        """
-        raise NotImplementedError('Metadata annotation needs to implement this method')
-
-    def build_kind_condition(self, kind_id, value, comparison='='):
-        return [BooleanComparison(kind_id, comparison, Value(value))]
-
-    def build_topic_condition(self, kind_id, comparison='='):
-        return self.build_kind_condition(kind_id, content_kinds.TOPIC, comparison)
-
-
-class DescendantCount(MetadataAnnotation):
-    def get_annotation(self, cte=None):
-        """
-        @see MPTTModel.get_descendant_count()
-        """
-        return Max(
-            Case(
-                # when selected node is topic, use algorithm to get descendant count
-                When(
-                    condition=WhenQ(*self.build_topic_condition(F('kind_id'))),
-                    then=(F('rght') - F('lft') - Value(1)) / Value(2)
-                ),
-                # integer division floors the result in postgres
-                default=Value(1)
-            )
-        )
-
-
-class DescendantAnnotation(MetadataAnnotation):
-    cte = TreeMetadataCTE
-    cte_columns = ('lft', 'rght')
-
-    def __init__(self, *args, **kwargs):
-        self.include_self = kwargs.pop('include_self', False)
-        super(DescendantAnnotation, self).__init__(*args, **kwargs)
-
-    def build_descendant_condition(self, cte):
-        """
-        @see MPTTModel.get_descendants()
-        """
-        left_op = '>='
-        right_op = '<='
-
-        if not self.include_self:
-            left_op = '>'
-            right_op = '<'
-
-        return [
-            BooleanComparison(cte.col.lft, left_op, F('lft')),
-            BooleanComparison(cte.col.lft, right_op, F('rght')),
-        ]
-
-
-class ResourceCount(DescendantAnnotation):
-    cte_columns = ('content_id', 'kind_id') + DescendantAnnotation.cte_columns
-
-    def get_annotation(self, cte):
-        resource_condition = self.build_topic_condition(F('kind_id'), '!=')
-
-        topic_condition = self.build_topic_condition(cte.col.kind_id, '!=')
-        topic_condition += self.build_descendant_condition(cte)
-
-        return Count(
-            Case(
-                # when selected node is not a topic, then count = 1
-                When(condition=WhenQ(*resource_condition), then=F('content_id')),
-                # when it is a topic, then count descendants
-                When(condition=WhenQ(*topic_condition), then=cte.col.content_id),
-                default=Value(None)
-            ),
-            distinct=True
-        )
-
-
-class CoachCount(DescendantAnnotation):
-    cte_columns = ('content_id', 'role_visibility') + DescendantAnnotation.cte_columns
-
-    def get_annotation(self, cte):
-        topic_condition = self.build_topic_condition(F('kind_id'))
-        topic_condition += self.build_descendant_condition(cte)
-        topic_condition += self.build_coach_condition(cte.col.role_visibility)
-
-        resource_condition = self.build_topic_condition(F('kind_id'), '!=')
-        resource_condition += self.build_coach_condition(F('role_visibility'))
-
-        return Count(
-            Case(
-                # when selected node is a coach topic, then count descendent content_id's
-                When(condition=WhenQ(*topic_condition), then=cte.col.content_id),
-                # when selected node is not a topic, count its content_id
-                When(condition=WhenQ(*resource_condition), then=F('content_id')),
-                default=Value(None)
-            ),
-            distinct=True
-        )
-
-    def build_coach_condition(self, role_visibility):
-        return [BooleanComparison(role_visibility, '=', Value(roles.COACH))]
-
-
-class HasChanged(DescendantAnnotation):
-    cte_columns = ('changed',) + DescendantAnnotation.cte_columns
-
-    def get_annotation(self, cte):
-        resource_condition = self.build_topic_condition(F('kind_id'), '!=')
-
-        whens = [
-            # when selected node is not a topic, just use its changed status
-            When(condition=WhenQ(*resource_condition), then=F('changed')),
-        ]
-
-        if self.include_self:
-            # when selected node is a topic and it's changed and including self, then return that
-            whens.append(When(condition=WhenQ(*[F('changed')]), then=F('changed')))
-
-        return Coalesce(
-            BoolOr(
-                Case(
-                    *whens,
-                    # fallback to aggregating descendant changed status when a unchanged topic
-                    default=cte.col.changed
-                )
-            ),
-            Value(False),
-            output_field=BooleanField()
-        )
-
-
-class SortOrderMax(DescendantAnnotation):
-    cte_columns = ('parent_id', 'sort_order') + DescendantAnnotation.cte_columns
-
-    def get_annotation(self, cte):
-        resource_condition = self.build_topic_condition(F('kind_id'), '!=')
-        topic_condition = self.build_child_condition(cte)
-
-        return Coalesce(
-            Max(
-                Case(
-                    # when selected node is not a topic, use its sort_order
-                    When(condition=WhenQ(*resource_condition), then=F('sort_order')),
-                    # when selected node is a topic, then find max of children
-                    When(condition=WhenQ(*topic_condition), then=cte.col.sort_order),
-                    default=Value(None)
-                )
-            ),
-            Value(1),
-            output_field=IntegerField()
-        )
-
-    def build_child_condition(self, cte):
-        return [
-            BooleanComparison(cte.col.parent_id, '=', F('id'))
-        ]
-
-
-class ResourceSize(DescendantAnnotation):
-    cte = FileMetadataCTE
-    cte_columns = ('file_size',)
-
-    def get_annotation(self, cte):
-        resource_condition = self.build_topic_condition(F('kind_id'), '!=')
-
-        return Sum(
-            Case(
-                # aggregate file_size when selected node is not a topic
-                When(
-                    condition=WhenQ(*resource_condition),
-                    then=Coalesce(cte.col.file_size, Value(0)),
-                ),
-                default=Value(0)
-            ),
-            output_field=IntegerField()
-        )
-
-
-class Metadata(object):
-    """
-    Helper class to query for various ContentNode metadata, for multiple node-trees, while
-    minimizing database query volume.
-
-    Example:
-        md = Metadata(['123...abc']).annotate(some_thing=MetadataAnnotation())
-        data = md.get('123...abc')
-    """
-    def __init__(self, node_pks, **annotations):
-        """
-        :param node_pks: list of ContentNode ID's
-        """
-        self.node_pks = node_pks
-        self.annotations = annotations
-        self.metadata = None
-
-    def annotate(self, **annotations):
-        """
-        :param annotations: Dict of annotations that should be instances of MetadataAnnotation
-        :return: A clone of Metadata with the new annotations
-        """
-        clone = Metadata(self.node_pks, **self.annotations)
-        clone.annotations.update(annotations)
-        return clone
-
-    def get(self, node_pk):
-        """
-        A dict of metadata for the node identified by `node_pk`
-        :param node_pk: An int
-        :return: A dict of metadata for the node identified by `node_pk`
-        """
-        if self.metadata is None:
-            self.retrieve_metadata()
-
-        return self.metadata.get(node_pk, None)
-
-    def retrieve_metadata(self):
-        """
-        :return: A dict of the metadata indexed by the node's ID
-        """
-        if len(self.annotations) == 0:
-            raise ValueError('No metadata to retrieve')
-
-        ctes = []
-
-        for field_name, annotation in self.annotations.items():
-            if not isinstance(annotation, MetadataAnnotation) or annotation.cte is None:
-                continue
-
-            if any(isinstance(cte, annotation.cte) for cte in ctes):
-                cte = next(cte for cte in ctes if isinstance(cte, annotation.cte))
-            else:
-                cte = annotation.cte(self.node_pks)
-                ctes.append(annotation.cte(self.node_pks))
-
-            if annotation.cte_columns:
-                cte.add_columns(list(annotation.cte_columns))
-
-        cte = None
-        query = ContentNode.objects.filter(pk__in=self.node_pks)
-        annotations = {}
-
-        if len(ctes) > 0:
-            for cte in ctes:
-                query = cte.join(query)
-                annotations.update({
-                    field_name: annotation.get_annotation(cte)
-                    for field_name, annotation in self.annotations.items()
-                    if isinstance(annotation, MetadataAnnotation)
-                    and annotation.cte and isinstance(cte, annotation.cte)
-                })
-
-        annotations.update({
-            field_name: annotation.get_annotation(cte)
-            for field_name, annotation in self.annotations.items()
-            if not isinstance(annotation, MetadataAnnotation) or annotation.cte is None
-        })
-
-        query = query.values('id').annotate(**annotations)
-
-        self.metadata = {}
-
-        # Finally, clear ordering (MPTT adds ordering by default)
-        for row in query.order_by():
-            self.metadata.update({row.pop('id'): row})
-
-        return self.metadata
